chore: refactor reporter events to a new file
