+++ reporter/index.js
@@ -1,10 +1,10 @@
-const utils = require('./lib/utils/index');
+const events = require('./lib/events.js');
 
 module.exports = function (newman, options, collectionOptions) {
-    newman.on('start', utils.handleStart);
-    newman.on('done', utils.handleDone);
+    newman.on('start', events.handleStart);
+    newman.on('done', events.handleDone);
 
-    newman.on('resume', utils.handleResume);
-    newman.on('pause', utils.handlePause);
-    newman.on('abort', utils.handleAbort);
+    newman.on('resume', events.handleResume);
+    newman.on('pause', events.handlePause);
+    newman.on('abort', events.handleAbort);
 };
